feat: Add cleanSet function for data manipulation
The `cleanSet` function is added to the `0x03-ES6_data_manipulation/8-clean_set.js` file. This function takes a Set object and a startString as parameters, filters the values that start with the given startString, slices the startString from each value, and joins the resulting values with a hyphen. This function enables cleaner data manipulation with Sets.